REFACTOR: Put creature class in own file
