+++ server/database/create.py
@@ -4,7 +4,7 @@ from sqlalchemy import create_engine
 from server.database.models import Base
 
 db_url = "sqlite:///server/database/instance/database.db"
-engine = create_engine(db_url)
+engine = create_engine(db_url, connect_args={"check_same_thread": False})
 
 Base.metadata.create_all(engine)
